Add network topology visualization script
This commit includes a Python script that visualizes network topology using IP addresses and their corresponding Autonomous System (AS) numbers. The script leverages the Graphviz library for creating the visualization and the IPInfo API to fetch AS information for given IP addresses. The resulting network topology diagram is saved as 'network_topology.gv'.